requirements.txt
asyncpg==0.29.0
aiohttp==3.9.1
python-dotenv==1.0.0
tqdm==4.66.1
